fix: prevent an infinite loop in measurementFieldSetChangeMgr (#25155)
The measurementFieldSetChangeMgr has a possibly infinite loop
if the writeRequests channel is closed while in the inner
loop to consolidate write requests. We need to check for ok
on channel receive and exit the loop when ok is false.

closes https://github.com/influxdata/influxdb/issues/25151

(cherry picked from commit 176fca2138a856ed54243a17d06e0926d200d382)

closes https://github.com/influxdata/influxdb/issues/25152