feat: delete cache method on LastCacheProvider
I also moved the write lock in the create_cache function for the
LastCacheProvider, to prevent a race condition between the time it was
getting the read lock to check the existence of the cache, and when it
gets the write lock to insert the new cache. Now, it just uses a single
write lock to do the check, and the subsequent insert.