feat: add compactor function for parquet files (#22)
This commit adds a new compaction function that takes in M files and
spits out N new files with a given limit. It sorts and dedupes based
off of the tags and time of the given parquet files. In the cases where
the time and tags are the same the data from the first file in the
passed in files is chosen.

Given two files with this tag set and a limit of two

1st: a, b, c, d, d
2nd: b, c, d

We'd create a record batch of:

a,b,b,c,c,d,d,d

which is 4 different series.

We will split it out into the following files:

1st: a
2nd: b, b
3rd: c, c
4th: d, d, d

In the first file the next series would go over the limit and it's a new
series so we close the file and start a new one. We only have 2 b for
the second one and 2 c for the 3rd one and so those are written out as
is. The 4th one however would "be over the limit", however we want data
for series to be colocated so we write out that entire series even if
that goes over the limit.

With this we'll be able to start work on the compaction scheduler so
that we can start compacting old data together which will help with
historical queries.

Closes #7

Note this commit also contains code synced from OSS.

---------

Co-authored-by: Paul Dix <paul@pauldix.net>
Co-authored-by: Trevor Hilton <thilton@influxdata.com>
Co-authored-by: Andrew Lamb <alamb@influxdata.com>