feat: track a snapshot sequence number
Snapshots get their own monotonically increasing sequence number independent
of the WAL file sequence number. This allows GET'ing the next snapshot
file from the object store using the most recent pulled snapshot, vs. having
to do a LIST operation to see if there are more snapshots.

The most recent snapshot sequence number is used to initialize the SnapshotTracker
on startup.

In addition, the catalog sequence number associated with a snapshot is tracked
on the SnapshotDetails.

Existing tests were updated to handle the struct changes and a new test
was added to the write buffer test suite that verifies when an existing
snapshot is loaded, the new snapshots generated henceforth use the correct
sequence number.