feat: introduce `WriteBufferPro` (#43)
* refactor: start on pro write buffer

* refactor: remove Persister trait in favour of concrete impl (#25260)

The Persister trait was only implemented by a single type, because the
underlying ObjectStore interface has several ways of being mocked, we
mock that instead of the Persister interface.

This commit removes the Persister trait, and moves its interface/impl
directly on a single Persister type in the persister module of the
influxdb3_write crate.

deny.toml had some incorrect field names in license.exceptions, those
were fixed from 'crate' to 'name'.

* refactor: have a working code split that compiles

The ReadWriteMode was shelled in to a working state, such that it could
be used in the main server, with all tests passing.

* refactor: extract catalog and last cache initialization

The initialization of catalog and last cache were being handled by the
write buffer implementation. This logic was moved out into a function on
the Persister type, so that the catalog and last cache are passed into
the write buffer. This was so that they can be initialized and shared
between the write buffer as well as replicas, or also so that they can
be initialized for a pure read replication setup, i.e., without a primary
buffer.

* feat: start on replicated buffer impl

* feat: enable initialization of replicated buffer

* feat: setup the replication interval

* feat: produce chunks from replicas and fetch WAL files repeatedly

* test: that a replica can replay a primary

Added a test that verifies that a replica can start, replay a
primary, and then replicate that primary on its replication interval.

This surfaced a few issues which were addressed in this commit.

Some changes in this PR will need to be upstreamed to OSS.

* docs: add missing doc line

* fix: ordering of snapshot clearing in replicated buffer