feat: enable read or read_write start modes
This enables two separate modes for the InfluxDB Pro buffer: read and
read_write. The latter is the same as the existing mode, while the former
is a pure read-only mode that can only replicate other hosts.

The read_write mode also allows replication, but is able to serve writes
as the current impliementation does, and therefore this is the default
mode of operation for the server.