feat: add FileIndex to compactor (#47)
This commit adds a FileIndex to the compactor. This lets us index
values for column names passed into the compactor function by being
able to look up the path for the name of the file.

Future work will need to be done for path rewriting in the index if
parquet files are moved to different paths. For now it provides us
an index of the files as they are written out into object storage.

This commit does a few things:
- The compactor now returns a struct with ParquetFile data, a FileIndex, and the necessary Object Store Paths from the compactor
- Provides a method to index a record batch passed into it
- Extends our compactor tests to include assertions for the indexes
- Changes the FileIndex to use a combined key for one lookup, rather
  than two lookups and two maps, simplifying it's logic

With these changes we can now create an index as we compact for
queries.