feat: implement compaction layout and planner (#42)
This adds a new crate called data_layout with the structures for tracking compacted data.

It also adds a planner to the compactor with the starting pieces to track snapshots coming from writing hosts and to generate plans to compact data from gen1 into gen2.

There's quite a bit here that is not yet fully fleshed out but contains everything for structure. Follow on PRs will combine this with the other compactor code and the new Pro code to get gen2 compaction running end to end.