fix: do not produce duplicate chunks with cache
This fixes a few tests that were broken after the change to populate the
in-memory parquet cache during snapshot. The write buffer will override
parquet chunks from the actual object store with those from the in-memory
cache if they are contained there.

Several tests needed to be switched to multi-threaded mode since the cache
was hooked up, as it uses a blocking task.