feat: Add u32 ID for Tables
This commit adds a u32 identifier for tables. This is mostly the same as
previous work done for the DbId and updates tests accordingly. The
important bit here is that I uncovered a double table creation bug
whereby we would create a new table, then insert the same new table as
part of the CatalogOp::AddFields, causing an off by one error for the id.
This commit removes the offending extra insert so that only one is done
to add the definition to the catalog.

Since we also needed to add the TableId to WriteBatch, we needed to do
custom serialization and deserialization logic as the tuple won't automatically
turn into a string for the serialization.