refactor: cache prune to use binary heap and timestamps
The cache was augmented to store hit_time on each entry, vs. a hit count.
This is updated each time an entry is requested from the cache, or when
the entry is first inserted.

The prune operation iterates over all entries, sorting their hit times and
sizes into a BinaryHeap, then pops elements off the heap until the target
free amount (currently hard-coded at 10%) is reached, to determine a cut-
off time.

The map is then pruned of all entries whose hit time is less than the
determined cut-off time.