fix: add host id to parquet file paths
Cleans up a legacy method for producing parquet file paths that was no
longer used in release code (was only being used in tests) and makes the
method that produces the parquet file path using a WAL sequence number
and i64 timestamp the defaul `new` constructor for the `ParquetFilePath`
type.

Tests that checked this function were modified accordingly.