feat: move Table Id/Name mapping into DB Schema
This is some follow up work to #25421 where we move the table mapping
into the DatabaseSchema struct itself. This code as a result has been
simplified compared to the original version and has the following
changes:

- By using the DatabaseSchema over the catalog in as many places as
  possible, we reduce the number of locks that we need to take on the
  Catalog. In many places we already had access to the schema and so
  this simplified things as we now would not need to look up tables
  with both the database id and the name/id of the table
- Code centered on making sure the ids was updated so that it only
  needed to occur when applying catalog ops
- We still need to add the id to name mapping in the validate lines fns
  for both v1 and v2, but it's now in the same spot as when we insert
  the table into the db schema making it more obvious what's going on

Overall, the code should be simpler to reason about and moving the map
into the DatabaseSchema was the right thing to do. The tests have also
been updated accordingly to handle these changes.

Closes #25429