refactor: Move ParquetFileId into influxdb3_id
This makes ParquetFileId consistent with our other ids and to avoid
confusion with them. It also fixes an issue where we would not
properly set the PersistedSnapshot `next_file_id` if we added a ParquetFile.
We instead now just always set the most available next id for the ParquetFileId,
every time to avoid confusion. A test has also been added to verify we have
the correct behavior.

Closes #25439