fix: added integration test for compaction
Adds an integration test for compaction. This uncovered a few bugs:

* PersistedSnapshot next_file_id is inaccurate: influxdata/influxdb#25439
* Metadata of the parquet file size returned by compaction function is wrong: #124

The first issue is worked around by assuming that file id is actually the last one used. But this should be investigated more in the OSS issue.

The second issue is fixed by doing a HEAD to the object store path to get the size. The existing Arrow writer API doesn't enable getting it.

Closes #124