feat: use hash values in file index
This removes the string interner from the file index. I realized that we don't actually need to keep the strings and instead could just use hash values of the column and value. This saves a bunch of memory space, avoids having a lock around a shared dictionary and is faster. A collision is possible, but in that case we'd simply get more Parquet file IDs back from the index than we otherwise would. The query will still work and return correct results.