feat: Add later generation compactions
This adds support for gen3 and gen4 compactions. I'm intentionally holding off on having it look for later generations than that as want to test this first.

This adds a new type of compaction plan and a planner function for the later generations. I've added a number of helper methods to build out the logic for creating the plans.

Closes #106