refactor: Simplify snapshot tracker and compaction planning
This removes the SnapshotTracker, opting instead to put snapshot data into the CompactedData structure.

This also removes the gen1_plan function, favoring the more generic function that can take a target output level.

This fixes an issue where previously compaction wouldn't run if a snapshot came in  and there were leftover gen1 files waiting. (before it would take two snapshots to trigger compaction).

This is work is a prerequisite to building out better compaction planning functionality, which will come in a follow on PR.