feat: Improve compaction planning and later compaction
This improves the compaction planner so that it purposefully leaves behind a younger generation so that it avoids having to re-compact later. It also picks up compactions of historical data and combines them into later generations.

This also updates the compaction runner so that it looks for snapshots to compact, but checks every 100ms to see if there are other later stage compactions to run. This is very useful in historical backfill situations where older compactions can back up.