refactor: working on last cache
The last cache needs a lot of work in its constructor, but the good news
is that I can likely rip out a lot of logic that relies on the Schema type
now, since the table/col definitions are more complete (we just hold the
schema type for iox_query/datafusion) and can be passed around more cheaply.

Likely need to pass that in to the cache creation so that the key col set
can be constructed more easily, then clean up logic that follows. The
cache does need to hold names, since it has code that works direct with
datafusion.