feat: add file index REST API
This commit changes quite a few things in order for us to have a file
index REST API. This API lets users set what columns to index on for
a given table when we run the compactor. By default we use the tag set
in order for us to do all of the sorting. With these changes a user
could specify tags and fields at the table, the database level, or both
that the compactor should sort and dedupe on. To accomplish this the
following changes were made:

- Introduce a new crate influxdb3_config which can be used for any
  configuration that does not need to be stored in the WAL. Instead we
  provide the ability to persist and retrieve it from the Object Store
  as needed. Currently this just contains the ProConfig which we can
  backport to OSS and add a regular config for use so that users don't
  have to keep specifying things like "hostname" etc.
- Adds a test suite to run the CLI and make sure it works as we would
  expect it too. This test mirrors the one added in the server
  integration tests to make sure that the CLI can do the same thing and
  behaves the same way when interacting with the server
- We thread the ProConfig throughout the codebase so that the compactor
  will either use the file index columns from the config or use the
  default with compactions

Closes #158